+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fix spacing on exception modal service point picker. Refs UICAL-145.
 * Use constant instead of hardcoded value for query limit. Refs UICAL-185.
 * Fix Calendar hours displayed off by one day in Regular hours Edit view. Refs UICAL-176.
+* Localize error message modal.  Refs UICAL-193.
 
 ## [7.0.2] (https://github.com/folio-org/ui-calendar/tree/v7.0.2) (2021-11-12)
 [Full Changelog](https://github.com/folio-org/ui-calendar/compare/v7.0.1...v7.0.2)